finished all controllers

+++ src/main/java/com/alvarorivas/finalproject/service/users/AccountHolderServiceImpl.java
@@ -3,7 +3,9 @@ package com.alvarorivas.finalproject.service.users;
 import com.alvarorivas.finalproject.model.users.AccountHolder;
 import com.alvarorivas.finalproject.repository.users.AccountHolderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
@@ -21,16 +23,44 @@ public class AccountHolderServiceImpl implements AccountHolderService{
 
     @Override
     public AccountHolder createAccHolder(AccountHolder accountHolder) {
-        return accountHolderRepository.save(accountHolder);
+
+        if (accountHolderRepository.findById(accountHolder.getAccountHolderId()).isPresent()){
+
+            return accountHolderRepository.save(accountHolder);
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User already exists");
+        }
     }
 
     @Override
-    public AccountHolder updateAccHolder(AccountHolder accountHolder) {
-        return accountHolderRepository.save(accountHolder);
+    public AccountHolder updateAccHolder(Integer id, AccountHolder accountHolder) {
+
+        Optional<AccountHolder> storedAccHolder = accountHolderRepository.findById(id);
+
+        if(storedAccHolder.isPresent()){
+
+            storedAccHolder.get().setBirthDate(accountHolder.getBirthDate());
+            storedAccHolder.get().setName(accountHolder.getName());
+            storedAccHolder.get().setMailingAddress(accountHolder.getMailingAddress());
+            storedAccHolder.get().setPrimaryAddress(accountHolder.getPrimaryAddress());
+
+            return accountHolderRepository.save(storedAccHolder.get());
+        }else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User ID not found");
+        }
     }
 
     @Override
     public void deleteAccHolder(Integer id) {
-        accountHolderRepository.deleteById(id);
+
+        Optional<AccountHolder> storedAccHolder = accountHolderRepository.findById(id);
+
+        if(storedAccHolder.isPresent()){
+
+            accountHolderRepository.deleteById(id);
+        }else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User ID not found");
+        }
+
     }
 }
